Make community pages loadable
